Revert "Create the database and all related functions"
