+++ python/error.py
@@ -1,0 +1,2 @@
+e404 = 'The requested page was not found on the server'
+e500 = 'Internal server error'
