+++ python/options.py
@@ -1,0 +1,18 @@
+food = {
+    #Food
+
+    #Pies
+    'foodPieScotch': 'Scotch Pie',
+    'foodPieSteak': 'Steak Pie',
+
+    #Burgers
+    'foodBurgerPlain': 'Plain Hamburger',
+    'foodBurgerCheese': 'Cheeseburger',
+
+    #Drinks
+
+    #Hot
+    'drinkHotBovril': 'Bovril'
+}
+
+options = [(value, label) for (value, label) in food.items()]
